change test due to change in test_case description; remove excess images
